requirements.txt
Pillow==5.4.1
pytest==4.3.1